Split flattener tests into smaller units
